+++ lib/src/tabbed_view_data.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/widgets.dart';
 import 'package:tabbed_view/src/tabbed_view.dart';
 import 'package:tabbed_view/src/tabbed_view_controller.dart';
+import 'package:tabbed_view/src/tabbed_view_menu_item.dart';
 
 /// Propagates parameters to internal widgets.
 class TabbedViewData {
@@ -14,7 +15,9 @@ class TabbedViewData {
       required this.selectToEnableButtons,
       this.closeButtonTooltip,
       this.tabsAreaButtonsBuilder,
-      this.draggableTabBuilder});
+      this.draggableTabBuilder,
+      required this.menuItems,
+      required this.menuItemsUpdater});
 
   final TabbedViewController controller;
   final bool contentClip;
@@ -26,4 +29,9 @@ class TabbedViewData {
   final String? closeButtonTooltip;
   final TabsAreaButtonsBuilder? tabsAreaButtonsBuilder;
   final DraggableTabBuilder? draggableTabBuilder;
+  final List<TabbedViewMenuItem> menuItems;
+  final MenuItemsUpdater menuItemsUpdater;
 }
+
+/// Updater for menu items
+typedef MenuItemsUpdater = void Function(List<TabbedViewMenuItem>);